requirements.txt
streamlit
scikit-learn
pandas
google-generativeai
python-dotenv
